requirements.txt
Django==3.0.8
djangorestframework==3.12.0
openpyxl==3.0.5